Adding log, reader and writer abstractions
