+++ hudi-utilities/src/test/java/org/apache/hudi/utilities/logger/InLineParquetWriter.java
@@ -1,0 +1,55 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.hudi.utilities.logger;
+
+import java.io.IOException;
+import java.util.List;
+import org.apache.avro.Schema;
+import org.apache.avro.generic.GenericRecord;
+import org.apache.hadoop.conf.Configuration;
+import org.apache.hadoop.fs.Path;
+import org.apache.parquet.avro.AvroParquetWriter;
+import org.apache.parquet.hadoop.metadata.CompressionCodecName;
+
+class InLineParquetWriter implements InLineWriter<GenericRecord> {
+
+  AvroParquetWriter parquetWriter;
+
+  InLineParquetWriter(Path path, Configuration conf, Schema schema) throws IOException {
+    parquetWriter = new AvroParquetWriter(path, schema,
+        CompressionCodecName.GZIP, 100 * 1024 * 1024, 1024 * 1024, true, conf);
+  }
+
+  @Override
+  public void write(GenericRecord record) throws IOException {
+    parquetWriter.write(record);
+  }
+
+  @Override
+  public void write(List<GenericRecord> records) throws IOException {
+    for (GenericRecord record : records) {
+      parquetWriter.write(record);
+    }
+  }
+
+  @Override
+  public void close() throws IOException {
+    parquetWriter.close();
+  }
+}
