+++ pom.xml
@@ -672,6 +672,11 @@
         <artifactId>hbase-client</artifactId>
         <version>${hbase.version}</version>
       </dependency>
+      <dependency>
+        <groupId>org.apache.hbase</groupId>
+        <artifactId>hbase-server</artifactId>
+        <version>${hbase.version}</version>
+      </dependency>
 
       <!-- Hive -->
       <dependency>
